requirements.txt
python-dotenv>=1.0.1
pdfplumber>=0.10.3
pdfminer.six>=20221105
pillow>=10.0.0
openai==0.28.1
PyGithub>=2.3.0
GitPython>=3.1.43
nbformat>=5.10.3
requests>=2.31.0
flask>=3.0.0
